feat: Add tile method to Schedule
The `tile` method has been added to the `Schedule` struct in order to support tiling of tensors. This method takes a tensor and a list of axes and performs tiling on the tensor based on the provided axes. The method also includes an assertion to ensure that the length of the axes is 2.

Refactor tensor.rs to add lazy loading for images
Refactor tensor.rs to remove redundant code and optimize performance
Refactor tensor.rs to use ToPrimeExpr trait for conversion in prod function