feat: Add `lowered` module to `hlir` namespace
This commit adds a new module called `lowered` to the `hlir` namespace. The `lowered` module contains code related to lowering the high-level intermediate representation (HLIR) to a lower-level representation. This module is responsible for performing transformations and optimizations on the HLIR.

The `lowered` module is added to the `hlir` namespace in the `tensor-compiler/src/lib.rs` file. This module will be used to implement the lowering process in future development.

Refactor tensor.rs to add lazy loading for images
Refactor tensor.rs to remove redundant code and optimize performance
Refactor tensor.rs to use ToPrimeExpr trait for conversion in prod function