Refactor `schedule.rs` to split and reorder tensor variables
This commit refactors the `schedule.rs` file in the `tensor-compiler` module. It introduces the `split` and `reorder` methods in the `Schedule` struct to handle tensor variable splitting and reordering. In the provided code changes, the `split` method is called twice on the `c` tensor variable, first with a split factor of 16 and then with a split factor of 7. Additionally, the `reorder` method is called to change the order of dimensions in the `c` tensor variable. These changes improve the scheduling and optimization capabilities of the `Schedule` module.

Refactor `iter.rs` to add support for stride and parent IterVar
Refactor `iter.rs` to add support for fused variables
Refactor `iter.rs` to use `IterVar` with additional fields
Refactor `temp.rs` to use `IterVar` with additional fields